Folder Restructure
Formalized folders into respective features w/ respective individual files
