+++ my-app/src/Components/Checkout/CartCheckout.css
@@ -1,35 +1,3 @@
-
-/*Homepage container styling*/
-.home-parent-container{
-    margin-top: 20px;
-}
-.home-display-title{
-    text-align: center;
-    font-family: Dialog;
-}
-
-/*Login & Signup Styling*/
-.form-title{
-    text-align: center;
-    margin-bottom: 10px;
-}
-.form-container{
-    border: 2px solid black;
-    border-radius: 5px;
-    padding: 20px;
-    margin-top: 25px;
-}
-
-/*ProductCard styling*/
-.card{
-    width: 20%;
-    background-color: antiquewhite;
-    text-align: center;
-    padding: 15px;
-    margin-right: 30px;
-    margin-top: 10px;
-}
-
 /*Cart styling*/
 .cart-parent-container{
     margin-top: 20px;
