+++ my-app/src/Components/Homepage/Home.css
@@ -1,0 +1,7 @@
+.home-parent-container{
+    margin-top: 20px;
+}
+.home-display-title{
+    text-align: center;
+    font-family: Dialog;
+}
