+++ my-app/src/Components/LoginSignup/LoginSignup.css
@@ -1,0 +1,13 @@
+.form-title{
+    text-align: center;
+    margin-bottom: 15px;
+    margin-top: 10px;
+}
+.form-container{
+    border: 2px solid black;
+    border-radius: 10px;
+    padding: 20px;
+    margin-top: 25px;
+    width: fit-content;
+;
+}
