+++ my-app/src/Components/Products/Products.css
@@ -1,0 +1,8 @@
+.card{
+    width: 20%;
+    background-color: #e9e6e1;
+    text-align: center;
+    padding: 15px;
+    margin-right: 30px;
+    margin-top: 10px;
+}
